Attempt at making use of VS Code's undo stack
Unfortunately, it doesn't seem possible.

The idea was to manually move a file to the trash and create a [`WorkspaceEdit`](https://code.visualstudio.com/api/references/vscode-api#WorkspaceEdit) to be able to undo it.

However, VS Code has separate undo stacks for the editor and the file explorer (https://code.visualstudio.com/updates/v1_52#_undo-file-operations-in-explorer), and `WorkspaceEdit` give access to only the editor one. There appears to be no way for extensions to access the undo stack of the file explorer.